Document get_remote_devices correction.
Previously, get_remote_devices returned a table of remote device types
to device numbers. However, device types are not unique, and device
numbers are unique only for each device type. So the function has been
altered to return a table with device types as keys and arrays of device
numbers as the values. This commit documents the corrected function.

Change-Id: I1a858a0093db83c146c29eadde2c533ac1a535ec