MDOC-82 Add webconfig and htaccess Documentation
The .webconfig and .htaccess file structures were documented in the
Designer help, however this was a bit odd as everything else you need
for custom web interface development is in the API docs.

Move this documentation to the API docs, and add Custom Groups which was
missing. This will be removed from the Designer help from 2.13 onwards.

Change-Id: I51ee99cca002f9edb5428b2a62fcd1de4305c2c6