MAG-15360 Document the group override target for Expert
Add the "group" override target to Expert's documentation for PUT
/api/override. Update the description for this endpoint to include
colour temperature.

This change also removes the global OVERRIDE_OPTIONS variable in favour
of hard-coded values. This is because the available override targets
differ between the PUT and DELETE endpoints causing too much
fragmentation for a central definition to be worthwhile.

Change-Id: Ifd32f4c994f975938adb00f874625c28598474ba